Reorganization of shapes in directory shapes
