+++ src/jujutracer.jl
@@ -39,8 +39,8 @@ export ImageTracer
 include("brdf.jl")
 export UniformPigment, CheckeredPigment, ImagePigment, AbstractBRDF, DiffusiveBRDF, SpecularBRDF, Material
 
-include("shapes.jl")
-export SurfacePoint, HitRecord, AbstractShape, AbstractSolid, Sphere, Box, Cylinder, Cone, Plane, Rectangle, Circle, World, ray_intersection, ray_intersection_list, Eval, boxed
+include("shapes/shapes.jl")
+#export SurfacePoint, HitRecord, AbstractShape, AbstractSolid, Sphere, Box, Cylinder, Cone, Plane, Rectangle, Circle, World, ray_intersection, ray_intersection_list, Eval, boxed
 
 include("meshes.jl")
 export Triangle, Parallelogram, ray_intersection, intersected, boxed
