+++ src/shapes/abstractshapes.jl
@@ -1,0 +1,17 @@
+#---------------------------------------------------------
+# Shapes
+#---------------------------------------------------------
+"""
+    abstract type AbstractShape
+
+Abstract type for all shapes. Not guaranteed to be water-tight. Cannot be used to create CSG shapes.
+"""
+abstract type AbstractShape end
+
+"""
+    abstract type AbstractSolid <: AbstractShape
+
+Abstract type for solid shapes. Considered water-tight. Can be used to create CSG shapes.
+Made concrete by [`Sphere`](@ref).
+"""
+abstract type AbstractSolid <: AbstractShape end
