+++ src/shapes/world.jl
@@ -1,0 +1,56 @@
+
+#---------------------------------------------------------
+# World type
+#---------------------------------------------------------
+"""
+    struct World
+
+A struct representing a collection of shapes in a 3D world.
+# Fields
+- `shapes::Vector{Shapes}`: the vector containing the shapes in the world.
+# Constructor
+- `World()`: creates a new `World` with an empty vector of shapes.
+- `World(S::Vector{Shapes})`: creates a new `World` with the specified vector of shapes.
+# See also
+- [`AbstractShape`](@ref): the abstract type for all shapes.
+- [`Sphere`](@ref): a concrete implementation of `AbstractShape` representing a sphere.
+- [`Plane`](@ref): a concrete implementation of `AbstractShape` representing a plane.
+"""
+struct World
+    shapes::Vector{AbstractShape}
+
+    function World()
+        new(Vector{AbstractShape}(nothing))
+    end
+
+    function World(S::Vector{AbstractShape})
+        new(S)
+    end
+end
+
+"""
+    ray_intersection(W::World, ray::Ray)
+
+Calculates the intersection of a ray with all shapes in the world.
+# Arguments
+- `W::World`: the world containing the shapes
+- `ray::Ray`: the ray to be intersected with the shapes in the worlds
+# Returns
+If there is an intersection, returns a `HitRecord` containing the hit information. Otherwise, returns `nothing`.
+"""
+function ray_intersection(W::World, ray::Ray)
+    dim = length(W.shapes)
+    closest = nothing
+
+    for i in 1:dim
+        inter = ray_intersection(W.shapes[i], ray)
+        if isnothing(inter)
+            continue
+        end
+        if (isnothing(closest) || inter.t < closest.t)
+            closest = inter
+        end
+    end
+
+    return closest
+end
